% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codingRegions.R
\name{codingRegions}
\alias{codingRegions}
\title{get gene coding regions (coding) grch37}
\usage{
codingRegions(genes)
}
\arguments{
\item{genes, }{a vctor list of hgnc gene names.}
}
\description{
This function extracts
}
\examples{
codingRegions()
}
\keyword{Glib}
