% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCons.R
\name{getCons}
\alias{getCons}
\title{A VCF read Function}
\usage{
getCons(Y, con)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function extracts
}
\examples{
varToChar()
}
\keyword{Glib}
