% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEffCons.R
\name{getEffCons}
\alias{getEffCons}
\title{This function runs through a list of snpEFF annotations and returns a index vector
of variants with the searched impact (LOW/HIGH/MODERATE/MODERATE)
if LOW is set to TRUE and a indexlist if LOW is FALSE}
\usage{
getEffCons(Y, con)
}
\arguments{
\item{Y}{a list of snpEFF annotations, con a string vector}
}
\description{
This function runs through a list of snpEFF annotations and returns a index vector
of variants with the searched impact (LOW/HIGH/MODERATE/MODERATE)
if LOW is set to TRUE and a indexlist if LOW is FALSE
}
\examples{
getEffCons(info(vcf)$EFF, cons)
getEffCons()
}
\keyword{Glib}
