% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdQqplot.R
\name{ggdQqplot}
\alias{ggdQqplot}
\title{QQplot of pvalues}
\usage{
ggdQqplot(pvector, main = NULL, ...)
}
\arguments{
\item{pvector, }{a vector of pvalues. main default NULL}
}
\description{
This function extracts
}
\examples{
ggdQqplot()
}
\keyword{Glib}
