\name{Glib-package}
\alias{Glib-package}
\alias{Glib}
\docType{package}
\title{
\packageTitle{Glib}
}
\description{
\packageDescription{Glib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Glib}
\packageIndices{Glib}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Glib}

Maintainer: \packageMaintainer{Glib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
