% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafFilter.R
\name{mafFilter}
\alias{mafFilter}
\title{A VCF read (by variantannotator) function to filter MAF}
\usage{
mafFilter(Y, maf, LOW = TRUE)
}
\arguments{
\item{a}{vector of MAFs,maf cutoff value, LOW=TRUE}
}
\description{
This function filter var over a thresgold returns a indexlist of variants under cutoff
if LOW is set to TRUE and a indexlist if LOW is FALSE
}
\examples{
mafFilter()
}
\keyword{Glib}
