% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varToChar.R
\name{varToChar}
\alias{varToChar}
\title{A Cat Function}
\usage{
varToChar(v1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you chenage variable names to a string
}
\examples{
varToChar()
}
\keyword{Glib}
